Update Demos: WebJS and NodeJS
Only demo simple demos for this repo.  More complex demos will move to another repo.
