+++ Bindings/WebAssembly/doxa.js
@@ -1,96 +1,101 @@
-(function (Doxa, Wasm) {
-	
-	var binarization;
-	
-	Doxa.getAlgorithms = function() {
-		let algorithms = new Array(10);
-		algorithms[Wasm.Binarization.Algorithms.OTSU.value] = { 
-			name: 'Otsu', 		params: {}, enum: Wasm.Binarization.Algorithms.OTSU,
-			date: 1979, title: 'A threshold selection method from gray-level histograms', authors: 'Nobuyuki Otsu'};
-		algorithms[Wasm.Binarization.Algorithms.BERNSEN.value] = { 
-			name: 'Bernsen', 	params: { window: 75, threshold: 100, 'contrast-limit': 25 }, enum: Wasm.Binarization.Algorithms.BERNSEN,
-			date: 1986, title: 'Dynamic thresholding of gray-level images', authors: 'John Bernsen'};
-		algorithms[Wasm.Binarization.Algorithms.NIBLACK.value] =	{
-			name: 'Niblack', 	params: { window: 223, k: -0.61 }, enum: Wasm.Binarization.Algorithms.NIBLACK,
-			date: 1986, title: 'An Introduction to Digital Image Processing', authors: 'Wayne Niblack'};
-		algorithms[Wasm.Binarization.Algorithms.SAUVOLA.value] =	{
-			name: 'Sauvola', 	params: { window: 26, k: 0.10 }, enum: Wasm.Binarization.Algorithms.SAUVOLA,
-			date: 1999, title: 'Adaptive document image binarization', authors: 'J. Sauvola, M. Pietikäinen'};
-		algorithms[Wasm.Binarization.Algorithms.WOLF.value] =	{
-			name: 'Wolf', 		params: { window: 20, k: 0.18 }, enum: Wasm.Binarization.Algorithms.WOLF,
-			date: 2003, title: 'Extraction and Recognition of Artificial Text in Multimedia Documents', authors: 'Christian Wolf, Jean-Michel Jolion'};
-		algorithms[Wasm.Binarization.Algorithms.NICK.value] =	{
-			name: 'N.I.C.K.', 	params: { window: 44, k:-0.10 }, enum: Wasm.Binarization.Algorithms.NICK,
-			date: 2009, title: 'Comparison of Niblack inspired Binarization methods for ancient documents', authors: 'Khurram Khurshid, Imran Siddiqi, Claudie Faure, Nicole Vincent'};
-		algorithms[Wasm.Binarization.Algorithms.SU.value] =	{
-			name: 'Su', 		params: { window: 26, k: 0.10 }, enum: Wasm.Binarization.Algorithms.SU,
-			date: 2010, title: 'Binarization of Historical Document Images Using the Local Maximum and Minimum', authors: 'Bolan Su, Shijian Lu, and Chew Lim Tan'};
-		algorithms[Wasm.Binarization.Algorithms.TRSINGH.value] =	{
-			name: 'T.R. Singh', params: { window: 75, k: 0.20 }, enum: Wasm.Binarization.Algorithms.TRSINGH,
-			date: 2011, title: 'A New local Adaptive Thresholding Technique in Binarization', authors: 'T. Romen Singh, Sudipta Roy, O. Imocha Singh, Tejmani Sinam, Kh. Manglem Singh'};
-		algorithms[Wasm.Binarization.Algorithms.BATAINEH.value] =	{
-			name: 'Bataineh', params: { }, enum: Wasm.Binarization.Algorithms.BATAINEH,
-			date: 2011, title: 'An adaptive local binarization method for document images based on a novel thresholding method and dynamic windows', authors: 'Bilal Bataineh, Siti Norul Huda Sheikh Abdullah, Khairuddin Omar'};
-		algorithms[Wasm.Binarization.Algorithms.ISAUVOLA.value] =	{
-			name: 'ISauvola', 	params: { window: 15, k: 0.01 }, enum: Wasm.Binarization.Algorithms.ISAUVOLA,
-			date: 2016, title: 'ISauvola: Improved Sauvola’s Algorithm for Document Image Binarization', authors: 'Zineb Hadjadj, Abdelkrimo Meziane, Yazid Cherfa, Mohamed Cheriet, Insaf Setitra'};
-		algorithms[Wasm.Binarization.Algorithms.WAN.value] =	{
-			name: 'Wan', 		params: { window: 75, k: 0.20 }, enum: Wasm.Binarization.Algorithms.WAN,
-			date: 2018, title: 'Binarization of Document Image Using Optimum Threshold Modification', authors: 'Wan Azani Mustafa, Mohamed Mydin M. Abdul Kader'};
+/**
+ * Doxa WASM
+ * A set of classes that further simplify the Doxa WASM interface.
+ * This same wrapper can be ran in NodeJS or directly in the web.
+ * See Demo/WebJs and Demo/NodeJS for an example of how to use it.
+ */
+
+const Doxa = {
+
+	Wasm: (typeof Module !== 'undefined') ? Module : require('./doxaWasm.js'),
+
+	initialize: async function() {
+		if (!Doxa.Wasm) throw new Error('Missing: Doxa WASM Module');
+
+		// Return the list of algorithms compiled into the WASM
+		const getAlgorithms = function() {
+			const algorithms = {}
+			for (const key in Doxa.Wasm.Binarization.Algorithms) {
+				const alg = Doxa.Wasm.Binarization.Algorithms[key];
+				if (alg.value === undefined) continue;
+
+				algorithms[key] = alg.value;
+			}
+			return algorithms;
+		}
+
+		return new Promise((resolve) => {
+			// Ensure the library has not already been initialized
+			if (typeof Doxa.Wasm.Binarization !== 'undefined') {
+				return resolve(getAlgorithms());
+			}
+
+			Doxa.Wasm.onRuntimeInitialized = async _ => {
+				resolve(getAlgorithms());
+			};
+		});
+	},
+
+	Image: class {
 		
-		return algorithms;
-	}
-	
-	Doxa.Image = function(width, height) {
-		"use strict";
-		let that = this;
-		this.width = 0;
-		this.height = 0;
-		this.size = 0;
-		
-		// Loads this Image object with an 8bit representation of an ImageData object
-		this.loadFromCanvas = function(imageData) {
-			initialize(imageData.width, imageData.height);
+		size = 0;
+
+		/**
+		 * Initialize an Image object.  All Image objects allocate WASM memory.
+		 * That memory should be freed, explicitly.  If you pass in a raw data array,
+		 * it will be converted to 8bit grayscale automatically.
+		 * @param {*} width Image Width
+		 * @param {*} height Image Height
+		 * @param {*} data Raw 32bit RGBA.  Optional.
+		 */
+		constructor (width, height, data) {
+			this.initialize(width, height);
 			
-			var data = new Uint8ClampedArray(Wasm.HEAPU8.buffer, this.heapPtr, this.size);
-			to8BitImage(data, imageData);
+			if (data) {
+				this.to8BitImage(this.data(), { width, height, data } );	
+			}
+		}
 			
-			return this;
-		};
-		
-		// Draws this Image directly to an HTML5 Canvas
-		this.draw = function(canvasId) {
-			var canvas = document.getElementById(canvasId);
+		/**
+		 * Draws the Image directly to an HTML5 Canvas.
+		 * The Canvas should accept 32bit data, which is standard.
+		 * @param {*} canvas HTML5 Canvas, or a node-canvas
+		 */
+		draw = function(canvas) {
 			canvas.width = this.width;
 			canvas.height = this.height;
 			
-            var ctx = canvas.getContext('2d');
-			var imageData = ctx.getImageData(0, 0, canvas.width, canvas.height);
+			const ctx = canvas.getContext('2d');
+			const imageData = ctx.getImageData(0, 0, canvas.width, canvas.height);
 						
-			var data = new Uint8ClampedArray(Wasm.HEAPU8.buffer, this.heapPtr, this.size);
-			from8BitImage(imageData, data);
+			this.from8BitImage(imageData, this.data());
 			
 			ctx.putImageData(imageData, 0, 0);
 		};
 		
-		// Frees the memory allocated by this object.
-		this.free = function() {
-			if (this.heapPtr != null) Wasm._free(this.heapPtr); 
+		data = function() {
+			return new Uint8ClampedArray(Doxa.Wasm.HEAPU8.buffer, this.heapPtr, this.size);
 		};
-		
-		// (re-)allocate memory for our image
-		function initialize(width, height) {
-			that.width = width;
-			that.height = height;
 			
-			if (width * height > that.size) {
-				that.free();
-				that.size = width * height;
-				that.heapPtr = Wasm._malloc(that.size);
+		/**
+		 * Frees the memory allocated by this object.
+		 */
+		free = function() {
+			if (this.heapPtr != null) Doxa.Wasm._free(this.heapPtr); 
+		};
+			
+		initialize(width, height) {
+			this.width = width;
+			this.height = height;
+			
+			if (width * height > this.size) {
+				this.free();
+				this.size = width * height;
+				this.heapPtr = Doxa.Wasm._malloc(this.size);
 			}
 		}
-		
-		function to8BitImage(imageBuffer, imageData) {
+			
+		to8BitImage(imageBuffer, imageData) {
 			const size32 = imageData.width * imageData.height * 4;
 
 			for (var idx = 0; idx < size32; idx += 4) {
@@ -98,12 +103,11 @@
 				const green = imageData.data[idx+1];
 				const blue = imageData.data[idx+2];
 
-				// TODO: Offer more grayscale options
 				imageBuffer[idx/4] = (red + green + blue) / 3;
 			}
 		}
 
-		function from8BitImage(imageData, imageBuffer) {
+		from8BitImage(imageData, imageBuffer) {
 			const size32 = imageData.width * imageData.height * 4;
 			
 			for (var idx = 0; idx < size32; idx += 4) {
@@ -114,24 +118,34 @@
 				imageData.data[idx+3] = 255;
 			}
 		}
-		
-		initialize(width, height);
-	};
+	},
 
-	Doxa.initialize = function(algorithm, image) {
-		if (binarization) binarization.delete();
-		
-		binarization = new Wasm.Binarization(algorithm);
-		binarization.initialize(image.heapPtr, image.width, image.height);
-	};
-	
-	Doxa.toBinary = function(image, parameters) {
-		let paramString = JSON.stringify(parameters);
 
-		binarization.toBinary(image.heapPtr, paramString);
-	};
-	
-	Doxa.calculatePerformance = function(groundTruth, image) {
-		return Wasm.calculatePerformance(groundTruth.heapPtr, image.heapPtr, image.width, image.height);
-	};
-}(window.Doxa = window.Doxa || {}, Module));
+	Binarization: class {
+
+		static toBinary = function(algorithm, imageIn, parameters, imageOut /* optional */) {
+			
+			// If no parameter is supplied, create an image object to return
+			if (!imageOut) {
+				imageOut = new Doxa.Image(imageIn.width, imageIn.height);
+			}
+			
+			const paramString = JSON.stringify(parameters);
+			const binarization = new Doxa.Wasm.Binarization(algorithm);
+			
+			binarization.initialize(imageIn.heapPtr, imageIn.width, imageIn.height);
+			binarization.toBinary(imageOut.heapPtr, paramString);
+			
+			return imageOut;
+		};
+		
+		static calculatePerformance = function(groundTruth, image) {
+			return Doxa.Wasm.calculatePerformance(groundTruth.heapPtr, image.heapPtr, image.width, image.height);
+		};
+	}
+}
+
+// Only export if running in NodeJS
+if (typeof module !== 'undefined') {
+	module.exports = { Doxa };
+}
