--- Demo/BinaryImageConverter/README.md
@@ -1,31 +1,0 @@
-# BinaryImageConverter v2.0
-## Introduction
-BinaryImageConverter is a demo application showing how to use the [Δoxa Binarization Framework](https://github.com/brandonmpetty/Doxa) to create binary images.
-
-**Algorithms**
-* Otsu - "A threshold selection method from gray-level histograms", 1979.
-* Bernsen - "Dynamic thresholding of gray-level images", 1986.
-* Niblack - "An Introduction to Digital Image Processing", 1986.
-* Sauvola - "Adaptive document image binarization", 1999.
-* Wolf - "Extraction and Recognition of Artificial Text in Multimedia Documents", 2003.
-* Gatos - "Adaptive degraded document image binarization", 2005. (Partial)
-* NICK - "Comparison of Niblack inspired Binarization methods for ancient documents", 2009.
-* Su - "Binarization of Historical Document Images Using the Local Maximum and Minimum", 2010.
-* T.R. Sing - "A New local Adaptive Thresholding Technique in Binarization", 2011.
-* WAN - "Binarization of Document Image Using Optimum Threshold Modification", 2018.
-
-**Native Image Support**
-* Portable Any-Map: PBM (P4), 8-bit PGM (P5), PPM (P6), PAM (P7)
-
-## Overview
-The goal of this application is to show how easily one can create binary images using ΔBF.
-You may specify one or more algorithms at a time, along with an optional Window Size and K-Value.
-
-## License
-CC0 - Brandon M. Petty, 2019
-
-To the extent possible under law, the author(s) have dedicated all copyright and related and neighboring rights to this software to the public domain worldwide. This software is distributed without any warranty.
-
-[View Online](https://creativecommons.org/publicdomain/zero/1.0/legalcode)
-
-"*Freely you have received; freely give.*" - Matt 10:8
