+++ Demo/NodeJS/.gitignore
@@ -1,0 +1,4 @@
+*.png
+doxa.js
+doxaWasm.js
+doxaWasm.wasm
