+++ Demo/NodeJS/package.json
@@ -4,7 +4,7 @@
   "description": "A simple demo showing how to use the Doxa library with NodeJS.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js"
   },
   "author": "",
   "license": "CC0-1.0",
