+++ Demo/NodeJS/setup.sh
@@ -1,0 +1,4 @@
+cp ../../Bindings/WebAssembly/dist/* .
+cp ../../Bindings/WebAssembly/doxa.js .
+cp ../../README/2JohnC1V3.png .
+cp ../../README/2JohnC1V3-GroundTruth.png .
